Desktop/email-automation/requirements.txt
bcrypt==4.0.1
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
sendgrid
passlib[bcrypt]
python-jose[cryptography]
authlib
google-auth
google-auth-oauthlib